OBJtoHKX.cs: Quote arguments to ProcessStartInfo
ParamManger.cs: Cast Const.FLVER_DMY_BOTTOM to int

SoundManager.cs: Parallelize BNK writing with Parallel.ForEach

+++ JortPob/Paramanager.cs
@@ -876,7 +876,7 @@ namespace JortPob
             int textId = textManager.AddActionButton(text);
 
             row["regionType"].Value.SetValue((byte)0); // cylinder
-            row["dummyPoly1"].Value.SetValue(Const.FLVER_DMY_BOTTOM); // area for entering door is at the bottom since the check is at the feet of the player character
+            row["dummyPoly1"].Value.SetValue((int)Const.FLVER_DMY_BOTTOM); // area for entering door is at the bottom since the check is at the feet of the player character
             row["dummyPoly2"].Value.SetValue(-1);
             row["radius"].Value.SetValue(width); // radius
             row["angle"].Value.SetValue(180); // angle from dmy
